remove shim. Lock in Lru-cache
